upgrade cobra_vendor to cobra 4.8 (issue #20)
Upgrade Cobra to version 4.8.

This is part of space-ros upgrade to Jazzy/Noble (Ubuntu24.04).

Tested with both Jazzy and Humble versions of space-ros:
* build works
* cobra tests run
* output of cobra is slightly different for MISRA ruleset due to message description copyright issues

Cobra 4.8 make install is now putting files directly to /usr/local/bin, so we have to use sudo.
Also removed the install part of cmake as bin_linux directory is not generated due to direct install.